extract formatMessageToParts, use 'literal' instead of KEY_LITERAL
