+++ __tests__/optional-format.test.js
@@ -1,5 +1,5 @@
 import { getIntl, text } from './helpers';
-import { formatDuration, formatDurationToParts, KEY_LITERAL } from '../index';
+import { formatDuration, formatDurationToParts } from '../index';
 
 describe('Optional format', () => {
   it('formats duration using hours (react)', () => {
@@ -39,93 +39,93 @@ describe('Optional format', () => {
 
     expect(formatDurationToParts(intl, 1, { format })).toEqual([
       { type: 'seconds', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'second' },
     ]);
 
     expect(formatDurationToParts(intl, 60, { format })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
     ]);
 
     expect(formatDurationToParts(intl, 61, { format })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'seconds', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'second' },
     ]);
 
     expect(formatDurationToParts(intl, 3600, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
     ]);
 
     expect(formatDurationToParts(intl, 3601, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'seconds', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'second' },
     ]);
 
     expect(formatDurationToParts(intl, 3602, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'seconds', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'seconds' },
     ]);
 
     expect(formatDurationToParts(intl, 3660, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
     ]);
 
     expect(formatDurationToParts(intl, 3661, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'seconds', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'second' },
     ]);
 
     expect(formatDurationToParts(intl, 7322, { format })).toEqual([
       { type: 'hours', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hours' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'minutes', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minutes' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'seconds', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'seconds' },
     ]);
 
     expect(formatDurationToParts(intl, 90000, { format })).toEqual([
       { type: 'hours', value: '25' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hours' },
     ]);
   });
@@ -170,93 +170,93 @@ describe('Optional format', () => {
     const format = '{days} {hours} {minutes}';
     expect(formatDurationToParts(intl, 1, { format })).toEqual([
       { type: 'minutes', value: '0' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minutes' },
     ]);
 
     expect(formatDurationToParts(intl, 60, { format })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
     ]);
 
     expect(formatDurationToParts(intl, 61, { format })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
     ]);
 
     expect(formatDurationToParts(intl, 110, { format })).toEqual([
       { type: 'minutes', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minutes' },
     ]);
 
     expect(formatDurationToParts(intl, 3600, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
     ]);
 
     expect(formatDurationToParts(intl, 3601, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
     ]);
 
     expect(formatDurationToParts(intl, 3660, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
     ]);
 
     expect(formatDurationToParts(intl, 3661, { format })).toEqual([
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minute' },
     ]);
 
     expect(formatDurationToParts(intl, 7322, { format })).toEqual([
       { type: 'hours', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hours' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'minutes', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'minutes' },
     ]);
 
     expect(formatDurationToParts(intl, 86400, { format })).toEqual([
       { type: 'days', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'day' },
     ]);
 
     expect(formatDurationToParts(intl, 90000, { format })).toEqual([
       { type: 'days', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'day' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'hours', value: '1' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hour' },
     ]);
 
     expect(formatDurationToParts(intl, 180000, { format })).toEqual([
       { type: 'days', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'days' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'hours', value: '2' },
-      { type: KEY_LITERAL, value: ' ' },
+      { type: 'literal', value: ' ' },
       { type: 'unit', value: 'hours' },
     ]);
   });
