+++ __tests__/timer-format.test.js
@@ -1,5 +1,5 @@
 import { getIntl, text } from './helpers';
-import { TIMER_FORMAT, formatDuration, formatDurationToParts, KEY_LITERAL } from '../index';
+import { TIMER_FORMAT, formatDuration, formatDurationToParts } from '../index';
 
 describe('Timer format', () => {
   it('formats duration in English with timer format (react)', () => {
@@ -33,55 +33,55 @@ describe('Timer format', () => {
 
     expect(formatDurationToParts(intl, 1, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '0' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '01' },
     ]);
 
     expect(formatDurationToParts(intl, 30, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '0' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '30' },
     ]);
 
     expect(formatDurationToParts(intl, 59, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '0' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '59' },
     ]);
 
     expect(formatDurationToParts(intl, 60, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '00' },
     ]);
 
     expect(formatDurationToParts(intl, 61, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '01' },
     ]);
 
     expect(formatDurationToParts(intl, 62, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '1' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '02' },
     ]);
 
     expect(formatDurationToParts(intl, 120, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '2' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '00' },
     ]);
 
     expect(formatDurationToParts(intl, 121, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '2' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '01' },
     ]);
 
     expect(formatDurationToParts(intl, 150, { format: TIMER_FORMAT })).toEqual([
       { type: 'minutes', value: '2' },
-      { type: KEY_LITERAL, value: ':' },
+      { type: 'literal', value: ':' },
       { type: 'seconds', value: '30' },
     ]);
   });
