+++ formatDuration.js
@@ -1,10 +1,5 @@
 import messages from './messages';
-
-/**
- * KEY_LITERAL is a symbol used as type for "literal" tokens.
- * It is a Symbol because "type" contains the name of the variable and could collide if it is a string.
- */
-export const KEY_LITERAL = typeof Symbol === 'function' ? Symbol('literal') : '@@literal@@';
+import { formatMessageToParts, KEY_LITERAL } from './formatMessageToParts';
 
 export const EXTENDED_FORMAT = 'EXTENDED_FORMAT';
 export const TIMER_FORMAT = 'TIMER_FORMAT';
@@ -81,7 +76,7 @@ export function formatDurationToParts(intl, seconds, { format } = {}) {
   const showFullMinutes = !hasHours && !hasDays;
   const showFullHours = !hasDays;
 
-  let tokenizedMessage = formatTokenizedMessage(intl, formattingMessage, {
+  let tokenizedMessage = formatMessageToParts(intl, formattingMessage, {
     days: hasDays ? valueFormatter({
       intl,
       value: fullDays,
@@ -163,6 +158,8 @@ export function formatDurationToParts(intl, seconds, { format } = {}) {
 
     const nextToken = tokenizedMessage[i + 1];
     if (nextToken.type !== KEY_LITERAL) {
+      token.type = 'literal';
+
       i++; // skip next token too
       continue;
     }
@@ -184,86 +181,9 @@ function formatUnit({ intl, value, showIfZero, message }) {
 
   const unit = intl.formatMessage(message, { value });
 
-  return formatTokenizedMessage(intl, messages.duration, { value, unit });
+  return formatMessageToParts(intl, messages.duration, { value, unit });
 }
 
 function padValue({ value, maxLengthIfPadded }) {
   return `0${value || '0'}`.substr(-maxLengthIfPadded);
-}
-
-/**
- * Like intl.formatMessage but returns an array of token split at the variables.
- * Similar to how <FormattedMessage> handles components in strings
- * https://github.com/yahoo/react-intl/blob/master/src/components/message.js#L58
- *
- * @returns Array
- */
-function formatTokenizedMessage(intl, message, values) {
-
-  // Note: Since both react-intl and we need a method like this, and has many many other use cases
-  // it'd be awesome if react-intl provided an intl.formatMessageToParts() natively instead of having this hack. (through https://github.com/yahoo/intl-messageformat)
-
-  const hasValues = values && Object.keys(values).length > 0;
-  if (!hasValues) {
-    return [{
-      type: 'literal',
-      value: intl.formatMessage(message, values),
-    }];
-  }
-
-  // Creates a token with a random UID that should not be guessable or
-  // conflict with other parts of the `message` string.
-  const uid = Math.floor(Math.random() * 0x10000000000).toString(16);
-
-  // Replace values variables with our tokens so we know where to inject the variables afterwards
-  const tokenizedValues = {};
-  Object.keys(values).forEach(name => {
-    tokenizedValues[name] = `@__TOKEN-${uid}-${name}__@`;
-  });
-
-  const formattedMessage = intl.formatMessage(message, tokenizedValues);
-
-  // Split the message into parts so the React Element values captured
-  // above can be inserted back into the rendered message.
-  const nodes = [];
-  const tokenMatcher = new RegExp(`@__TOKEN-${uid}-(.+?)__@`);
-
-  let str = formattedMessage;
-  while (str) {
-    // find first token
-    const match = str.match(tokenMatcher);
-    if (!match) {
-      // no token left, add everything else as literal
-      nodes.push({
-        type: KEY_LITERAL,
-        value: str,
-      });
-
-      break;
-    }
-
-    const matchStart = match.index;
-    const matchLength = match[0].length;
-    if (matchStart !== 0) {
-      // add all chars before the matched token as a literal node
-      const strStart = str.substr(0, matchStart);
-
-      nodes.push({
-        type: KEY_LITERAL,
-        value: strStart,
-      });
-    }
-
-    // add the matched token as variable node
-    const valueKey = match[1];
-    nodes.push({
-      type: valueKey,
-      value: values[valueKey],
-    });
-
-    // remove all chars that have been added to `nodes` during this iteration
-    str = str.substr(matchStart + matchLength);
-  }
-
-  return nodes;
 }
